virt-controller-vmi: move datavolume event handlers to datavolume
